requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.2.0
drf-yasg==1.21.7
